+++ Tests/Microsoft.AppCenter.Test.Functional/Distribute/DistributeUpdateTest.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
 using System.Threading.Tasks;
 using Xunit;
 
@@ -9,9 +8,12 @@ namespace Microsoft.AppCenter.Test.Functional.Distribute
 {
     using Distribute = Microsoft.AppCenter.Distribute.Distribute;
     using AppCenter = Microsoft.AppCenter.AppCenter;
+    using UpdateTrack = Microsoft.AppCenter.Distribute.UpdateTrack;
 
     public enum DistributeTestType
     {
+        SaveMockUpdateToken,
+        EnableDebuggableBuilds,
         FreshInstallAsync,
         CheckUpdateAsync,
         OnResumeActivity,
@@ -28,11 +30,16 @@ namespace Microsoft.AppCenter.Test.Functional.Distribute
         public DistributeUpdateTest()
         {
             Utils.DeleteDatabase();
+            Distribute.UnsetInstance();
+            DistributeEvent?.Invoke(this, DistributeTestType.Clear);
         }
 
         [Fact]
         public async Task GetLastReleaseDetailsAsync()
         {
+            // Enable Distribute for debuggable builds.
+            DistributeEvent?.Invoke(this, DistributeTestType.EnableDebuggableBuilds);
+
             // Save data to preference.
             DistributeEvent?.Invoke(this, DistributeTestType.CheckUpdateAsync);
 
@@ -48,8 +55,7 @@ namespace Microsoft.AppCenter.Test.Functional.Distribute
             AppCenter.Start(Config.ResolveAppSecret(), typeof(Distribute));
 
             // Wait for "startService" log to be sent.
-            Task.WaitAny(startServiceTask, Task.Delay(5000));
-
+            await startServiceTask;
             DistributeEvent?.Invoke(this, DistributeTestType.OnResumeActivity);
 
             // Wait when Distribute will start.
@@ -61,6 +67,88 @@ namespace Microsoft.AppCenter.Test.Functional.Distribute
             // Verify response.
             Assert.Equal("GET", result.Method);
             Assert.True(result.Uri.Contains("releases/latest?release_hash"));
+            Assert.True(result.Uri.Contains(Config.ResolveAppSecret()));
+
+            // Clear.
+            DistributeEvent?.Invoke(this, DistributeTestType.Clear);
+        }
+
+        [Fact]
+        public async Task SetUpdateTrackPublicTest()
+        {
+            // Enable Distribute for debuggable builds.
+            DistributeEvent?.Invoke(this, DistributeTestType.EnableDebuggableBuilds);
+
+            // Setup network adapter.
+            var httpNetworkAdapter = new HttpNetworkAdapter();
+            DependencyConfiguration.HttpNetworkAdapter = httpNetworkAdapter;
+            var eventTask = httpNetworkAdapter.MockRequest(request => request.Method == "GET");
+            var startServiceTask = httpNetworkAdapter.MockRequestByLogType("startService");
+
+            // Start AppCenter.
+            AppCenter.UnsetInstance();
+            AppCenter.LogLevel = LogLevel.Verbose;
+            Distribute.UpdateTrack = UpdateTrack.Public;
+            AppCenter.Start(Config.ResolveAppSecret(), typeof(Distribute));
+
+            // Wait for "startService" log to be sent.
+            await startServiceTask;
+            DistributeEvent?.Invoke(this, DistributeTestType.OnResumeActivity);
+
+            // Wait when Distribute will start.
+            await Distribute.IsEnabledAsync();
+
+            // Wait for processing event.
+            var result = await eventTask;
+
+            // Verify response.
+            Assert.Equal("GET", result.Method);
+            Assert.True(result.Uri.Contains("public"));
+            Assert.True(result.Uri.Contains("releases/latest"));
+            Assert.True(result.Uri.Contains(Config.ResolveAppSecret()));
+
+            // Clear.
+            DistributeEvent?.Invoke(this, DistributeTestType.Clear);
+        }
+
+        [Fact]
+        public async Task SetUpdateTrackPrivateTest()
+        {
+            // Enable Distribute for debuggable builds.
+            DistributeEvent?.Invoke(this, DistributeTestType.EnableDebuggableBuilds);
+
+            // Save data to preference.
+            DistributeEvent?.Invoke(this, DistributeTestType.CheckUpdateAsync);
+
+            // Setup network adapter.
+            var httpNetworkAdapter = new HttpNetworkAdapter();
+            DependencyConfiguration.HttpNetworkAdapter = httpNetworkAdapter;
+            var eventTask = httpNetworkAdapter.MockRequest(request => request.Method == "GET");
+            var startServiceTask = httpNetworkAdapter.MockRequestByLogType("startService");
+
+            // Start AppCenter.
+            AppCenter.UnsetInstance();
+            AppCenter.LogLevel = LogLevel.Verbose;
+            Distribute.UpdateTrack = UpdateTrack.Private;
+
+            // MockUpdateToken.
+            DistributeEvent?.Invoke(this, DistributeTestType.SaveMockUpdateToken);
+            AppCenter.Start(Config.ResolveAppSecret(), typeof(Distribute));
+
+            // Wait for "startService" log to be sent.
+            await startServiceTask;
+            DistributeEvent?.Invoke(this, DistributeTestType.OnResumeActivity);
+
+            // Wait when Distribute will start.
+            await Distribute.IsEnabledAsync();
+
+            // Wait for processing event.
+            var result = await eventTask;
+
+            // Verify response.
+            Assert.Equal("GET", result.Method);
+            Assert.False(result.Uri.Contains("public"));
+            Assert.True(result.Uri.Contains("releases/private/latest?release_hash"));
             Assert.True(result.Uri.Contains(Config.ResolveAppSecret()));
 
             // Clear.